Fixing SEAMFACES-126
Logging a warning instead of an exception. This is really a
configuration issue. A 401 is actually the correct thing to do if there
is no configured page.

Also wrapping any output in an if statement to make sure we don't send
any of our output if the users have already pushed things out to the
output stream.